requirements.txt
acres==0.5.0
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.3
ci-info==0.3.0
click==8.2.1
colorama==0.4.6
comm==0.2.2
configobj==5.0.9
configparser==7.2.0
cryptography==45.0.6
debugpy==1.8.12
decorator==5.1.1
distlib==0.4.0
distro==1.9.0
dotenv==0.9.9
etelemetry==0.3.1
executing==2.2.0
fastapi==0.116.1
filelock==3.19.1
Flask==3.1.2
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-storage==3.3.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
identify==2.6.13
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
looseversion==1.3.0
lxml==6.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
nodeenv==1.9.1
numpy==2.2.4
openai==1.99.9
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathlib==1.0.1
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
platformdirs==4.3.6
pre_commit==4.3.0
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==6.32.0
prov==2.1.1
psutil==6.1.1
pure_eval==0.2.3
puremagic==1.30
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.1
Pygments==2.19.1
PyMuPDF==1.26.3
pyparsing==3.2.3
pypdf==6.0.0
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq==26.2.1
rdflib==7.1.4
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
rpds-py==0.27.0
rsa==4.9.1
scipy==1.16.1
setuptools==80.9.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.47.2
tenacity==9.1.2
tiktoken==0.11.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
virtualenv==20.34.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
gunicorn